matchingapp/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
